start my project
Hello to everyone who is reading these details. Today, May 2, 1403, I started my personal website project and I plan to create this website for resume and practice. I also plan to provide the possibility of real shopping for my website and add a shopping cart to this template. I hope it won't take too long and the project will be successful. In addition, my ready template has a front bug and I have to solve it with the help of my friends or by myself.

Co-Authored-By: Mohammad Mahdi Niknam <141505892+niknam1382@users.noreply.github.com>